cli: add magic-attach flow support
Add support for the magic-attach flow through the attach CLI command.
This flow allows users to attach to Ubuntu Pro subscription
without a token. When using that command, the Pro client will
output a user code on the screen and the user will then login
into the Pro portal url and input that token there. Once that is
done, our code will detect that and fetch the token directly
from the Contract Server. Once we have the token, we can then
attach as usual